mindsdb/integrations/handlers/magento_handler/requirements.txt
my-magento